backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
httpx
pytest
pytest-asyncio
docker
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
